feat(configuration.ts): add ability to set user agent
This commit adds the ability to set the user agent for the OpenAI API client configuration. A user agent name and version can now be specified in the configuration parameters, which will be included in the headers sent to the server. The default user agent name is 'OpenAI' and the default version is '3.2.1'. The user agent can be accessed and modified using the `userAgent` getter and setter functions of the `Configuration` class.