feat: add user agent name and version to OpenAIApi config
Add user agent name and version to the OpenAIApi configuration in
soakp-client.ts. This will allow for better tracking of client usage on
the OpenAI platform.

Also import fs in soakp-client.ts to read the version number from
package.json and use it as the user agent version.

Finally, add a script tag to index.html to load the main.js file.