Merge develop (#1)
* feat(configuration.ts): import packageJson and FormData

- Import packageJson from './package.json' and import * as FormData from 'form-data' in configuration.ts. This will allow a better use of these modules in the code.
update dependencies and TS config
- Update package.json dependencies.
- Update compilerOptions in tsconfig.json to target ES2018 and use esnext module.
- Add strict, esModuleInterop, and allowSyntheticDefaultImports options.
- Update lib, include, and exclude options.
- Add watchOptions to exclude certain directories and files.

* refactor: Directory structure

* refactor:  Moved files

* feat: Add demo HTML root

* feat: Add Docker nginx configuration

* Added public HTML template folder for snowpack

* feat: Add various configs

* feat: Added static page template for demo

* feat: Add static demo page

* refactor(tsconfig): exclude demo folder from typescript compilation

refactor(snowpack): update snowpack config to use public folder as root and exclude .snowpack from build process. Also update known entrypoints for form-data and axios.

* refactor: Update demo

* chore: Update documentation

* feat: Update build configuration

* feat: update demo

* feat(configuration.ts): add ability to set user agent

This commit adds the ability to set the user agent for the OpenAI API client configuration. A user agent name and version can now be specified in the configuration parameters, which will be included in the headers sent to the server. The default user agent name is 'OpenAI' and the default version is '3.2.1'. The user agent can be accessed and modified using the `userAgent` getter and setter functions of the `Configuration` class.

* feat: add user agent name and version to OpenAIApi config

Add user agent name and version to the OpenAIApi configuration in
soakp-client.ts. This will allow for better tracking of client usage on
the OpenAI platform.

Also import fs in soakp-client.ts to read the version number from
package.json and use it as the user agent version.

Finally, add a script tag to index.html to load the main.js file.

* build: remove commented out external dependencies from packageOptions in snowpack.config.js

* chore: Update SOAKP client and server documentation links in README.md

* chore: Removed unnecessary files

* chore:  Updated demo

* feat(soakp-client): update user agent version to 1.0.1

* feat: update package.json main entry

Update the main entry in package.json to point to the demo folder instead of the dist folder. The new main entry is "./demo/js/soakp-client.js".