feat: Update UA version in SoakpClient constructor
Update the User-Agent (UA) version in the constructor of SoakpClient
to "1.0.1". The previous version was being read from the package.json
file using fs, which is unnecessary and potentially problematic in a
browser environment. This commit updates the UA version to a static
string, which is a more reliable and efficient approach.

- Import packageJson from './package.json' and import * as FormData from 'form-data' in configuration.ts. This will allow a better use of these modules in the code.
update dependencies and TS config
- Update package.json dependencies.
- Update compilerOptions in tsconfig.json to target ES2018 and use esnext module.
- Add strict, esModuleInterop, and allowSyntheticDefaultImports options.
- Update lib, include, and exclude options.
- Add watchOptions to exclude certain directories and files.

This commit adds the ability to set the user agent for the OpenAI API client configuration. A user agent name and version can now be specified in the configuration parameters, which will be included in the headers sent to the server. The default user agent name is 'OpenAI' and the default version is '3.2.1'. The user agent can be accessed and modified using the `userAgent` getter and setter functions of the `Configuration` class.

Add user agent name and version to the OpenAIApi configuration in
soakp-client.ts. This will allow for better tracking of client usage on
the OpenAI platform.

Also import fs in soakp-client.ts to read the version number from
package.json and use it as the user agent version.

Finally, add a script tag to index.html to load the main.js file.

Update the main entry in package.json to point to the demo folder instead of the dist folder. The new main entry is "./demo/js/soakp-client.js".

Export SoakpClient as default module to remove the need to destructure it
when importing into other modules. Remove unused fs import from soakp-client.ts.

Add a test suite for SoakpClient with tests for constructor and makeRequest
methods. The tests cover valid and invalid scenarios for createCompletion,
createChatCompletion, createEdit and invalid endpoint. The tests also mock
console.log and console.error methods to ensure they are called as expected.
This commit adds support for additional endpoints in the `makeRequest()` method of the SoakpClient class. Specifically, it adds support for the `/completion`, `/chat/completion`, and `/edits` endpoints. The commit also includes updated tests to ensure that the new endpoints are functioning correctly.

Added Jest configuration file to the project with basic setup for TypeScript and node test environment. Also added coverage report setup with text-summary and lcov reporters.

Update parser options in .eslintrc.json and exclude .js files in tsconfig.json