feat: Update UA version in SoakpClient constructor
Update the User-Agent (UA) version in the constructor of SoakpClient
to "1.0.1". The previous version was being read from the package.json
file using fs, which is unnecessary and potentially problematic in a
browser environment. This commit updates the UA version to a static
string, which is a more reliable and efficient approach.