Develop (#3)
* feat(configuration.ts): import packageJson and FormData

- Import packageJson from './package.json' and import * as FormData from 'form-data' in configuration.ts. This will allow a better use of these modules in the code.
update dependencies and TS config
- Update package.json dependencies.
- Update compilerOptions in tsconfig.json to target ES2018 and use esnext module.
- Add strict, esModuleInterop, and allowSyntheticDefaultImports options.
- Update lib, include, and exclude options.
- Add watchOptions to exclude certain directories and files.

* refactor: Directory structure

* refactor:  Moved files

* feat: Add demo HTML root

* feat: Add Docker nginx configuration

* Added public HTML template folder for snowpack

* feat: Add various configs

* feat: Added static page template for demo

* feat: Add static demo page

* refactor(tsconfig): exclude demo folder from typescript compilation

refactor(snowpack): update snowpack config to use public folder as root and exclude .snowpack from build process. Also update known entrypoints for form-data and axios.

* refactor: Update demo

* chore: Update documentation

* feat: Update build configuration

* feat: update demo

* feat(configuration.ts): add ability to set user agent

This commit adds the ability to set the user agent for the OpenAI API client configuration. A user agent name and version can now be specified in the configuration parameters, which will be included in the headers sent to the server. The default user agent name is 'OpenAI' and the default version is '3.2.1'. The user agent can be accessed and modified using the `userAgent` getter and setter functions of the `Configuration` class.

* feat: add user agent name and version to OpenAIApi config

Add user agent name and version to the OpenAIApi configuration in
soakp-client.ts. This will allow for better tracking of client usage on
the OpenAI platform.

Also import fs in soakp-client.ts to read the version number from
package.json and use it as the user agent version.

Finally, add a script tag to index.html to load the main.js file.

* build: remove commented out external dependencies from packageOptions in snowpack.config.js

* chore: Update SOAKP client and server documentation links in README.md

* chore: Removed unnecessary files

* chore:  Updated demo

* feat(soakp-client): update user agent version to 1.0.1

* feat: update package.json main entry

Update the main entry in package.json to point to the demo folder instead of the dist folder. The new main entry is "./demo/js/soakp-client.js".

* feat: export SoakpClient as default module

Export SoakpClient as default module to remove the need to destructure it
when importing into other modules. Remove unused fs import from soakp-client.ts.

* feat: Added tests configurations

* feat: add SoakpClient test suite

Add a test suite for SoakpClient with tests for constructor and makeRequest
methods. The tests cover valid and invalid scenarios for createCompletion,
createChatCompletion, createEdit and invalid endpoint. The tests also mock
console.log and console.error methods to ensure they are called as expected.
This commit adds support for additional endpoints in the `makeRequest()` method of the SoakpClient class. Specifically, it adds support for the `/completion`, `/chat/completion`, and `/edits` endpoints. The commit also includes updated tests to ensure that the new endpoints are functioning correctly.

* 1.1.0

* build: add Jest configuration file

Added Jest configuration file to the project with basic setup for TypeScript and node test environment. Also added coverage report setup with text-summary and lcov reporters.

Update parser options in .eslintrc.json and exclude .js files in tsconfig.json

* feat: add GitHub Actions workflows for remote deployment and its simulation

Added two new GitHub Actions workflows for remote deployment: one for dry run
and one for actual deployment. The workflows run on the `master` branch and
publish the package to both GitHub Packages and NPM on a release event.

The `.github/workflows/npm-publish.yml` workflow handles actual deployment,
while the `.github/workflows/npm-publish-dry-run.yml` workflow performs a dry
run of the deployment. The `dependabot.yml` file was also added to support
version updates.

All workflows use Node.js version 16 and run on Ubuntu latest. The package is
built in production mode and published using the appropriate registry URL and
authentication token.

* feat: add browser usage documentation and update README

* Update test.yml

* feat: add configuration for remote theme, plugins, and Google Analytics

Add configuration for the remote theme, plugins, and Google Analytics in
the `_config.yml` file. The remote theme is set to
`pages-themes/midnight@v0.2.0`, the `jekyll-remote-theme` plugin is added
to the plugins list, `show_downloads` is set to true.

* feat: Update UA version in SoakpClient constructor

Update the User-Agent (UA) version in the constructor of SoakpClient
to "1.0.1". The previous version was being read from the package.json
file using fs, which is unnecessary and potentially problematic in a
browser environment. This commit updates the UA version to a static
string, which is a more reliable and efficient approach.